[API] Refactor StateProviders
