+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/AdminUserPassword.xml
@@ -60,7 +60,7 @@
                 class="ApiPlatform\Metadata\Patch"
                 uriTemplate="/admin/administrators/reset-password/{token}"
                 input="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword"
-                provider="sylius_api.state_provider.admin.account.reset_password.item"
+                provider="sylius_api.state_provider.admin.admin_user.reset_password"
             >
                 <denormalizationContext>
                     <values>
