+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Command/VerifyShopUser.xml
@@ -61,7 +61,7 @@
                 class="ApiPlatform\Metadata\Patch"
                 uriTemplate="/shop/customers/verify/{token}"
                 input="Sylius\Bundle\ApiBundle\Command\Account\VerifyShopUser"
-                provider="sylius_api.state_provider.shop.account.verify_shop_user.item"
+                provider="sylius_api.state_provider.shop.shop_user.verify_shop_user"
             >
                 <denormalizationContext>
                     <values>
