+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/ShopUserResetPassword.xml
@@ -61,7 +61,7 @@
                 output="false"
                 status="202"
                 uriTemplate="/shop/customers/reset-password/{token}"
-                provider="sylius_api.state_provider.shop.account.reset_password.item"
+                provider="sylius_api.state_provider.shop.shop_user.reset_password"
             >
                 <denormalizationContext>
                     <values>
