+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/state_providers.xml
@@ -41,17 +41,17 @@
             <tag name="api_platform.state_provider" priority="10"/>
         </service>
 
-        <service id="sylius_api.state_provider.shop.account.reset_password.item" class="Sylius\Bundle\ApiBundle\StateProvider\Shop\Account\ResetPassword\ItemProvider">
+        <service id="sylius_api.state_provider.shop.shop_user.reset_password" class="Sylius\Bundle\ApiBundle\StateProvider\Shop\ShopUser\ResetPasswordProvider">
             <argument type="service" id="Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface" />
             <tag name="api_platform.state_provider" priority="10" />
         </service>
 
-        <service id="sylius_api.state_provider.shop.account.verify_shop_user.item" class="Sylius\Bundle\ApiBundle\StateProvider\Shop\Account\VerifyShopUser\ItemProvider">
+        <service id="sylius_api.state_provider.shop.shop_user.verify_shop_user" class="Sylius\Bundle\ApiBundle\StateProvider\Shop\ShopUser\VerifyShopUserProvider">
             <argument type="service" id="Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface" />
             <tag name="api_platform.state_provider" priority="10" />
         </service>
 
-        <service id="sylius_api.state_provider.admin.account.reset_password.item" class="Sylius\Bundle\ApiBundle\StateProvider\Admin\Account\ResetPassword\ItemProvider">
+        <service id="sylius_api.state_provider.admin.admin_user.reset_password" class="Sylius\Bundle\ApiBundle\StateProvider\Admin\AdminUser\ResetPasswordProvider">
             <argument type="service" id="Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface" />
             <tag name="api_platform.state_provider" priority="10" />
         </service>
