+++ src/Sylius/Bundle/ApiBundle/StateProvider/Shop/ShopUser/ResetPasswordProvider.php
@@ -11,18 +11,18 @@
 
 declare(strict_types=1);
 
-namespace Sylius\Bundle\ApiBundle\StateProvider\Admin\Account\ResetPassword;
+namespace Sylius\Bundle\ApiBundle\StateProvider\Shop\ShopUser;
 
 use ApiPlatform\Metadata\Operation;
 use ApiPlatform\Metadata\Patch;
 use ApiPlatform\State\ProviderInterface;
-use Sylius\Bundle\ApiBundle\SectionResolver\AdminApiSection;
-use Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword;
+use Sylius\Bundle\ApiBundle\Command\Account\ResetPassword;
+use Sylius\Bundle\ApiBundle\SectionResolver\ShopApiSection;
 use Sylius\Bundle\CoreBundle\SectionResolver\SectionProviderInterface;
 use Webmozart\Assert\Assert;
 
 /** @implements ProviderInterface<ResetPassword> */
-final readonly class ItemProvider implements ProviderInterface
+final readonly class ResetPasswordProvider implements ProviderInterface
 {
     public function __construct(private SectionProviderInterface $sectionProvider)
     {
@@ -32,7 +32,7 @@ final readonly class ItemProvider implements ProviderInterface
     {
         Assert::true(is_a($operation->getClass(), ResetPassword::class, true));
         Assert::isInstanceOf($operation, Patch::class);
-        Assert::isInstanceOf($this->sectionProvider->getSection(), AdminApiSection::class);
+        Assert::isInstanceOf($this->sectionProvider->getSection(), ShopApiSection::class);
         Assert::stringNotEmpty($uriVariables['token'] ?? null, 'Token is required.');
 
         return new ResetPassword($uriVariables['token']);
